Removed the native NostalgicPlayerLibrary.
Made some test to see if C# versions of the mixer is faster/slower/equal to the C versions. When calling the C versions, there is some overhead by calling unmanaged code. In fact, it turns out that the C# versions are faster, but we're talking nanoseconds here, so not much. So I decided, to get rid of the C code, so the main player is now in pure C#.

The main reason I have the C code to begin with, is because the source byte array could either be 8 or 16 bit samples, and I didn't want to make a copy/conversion of the array into a short array when 16 bit is needed (too slow). But I found the Span<> class, which can be used to cast an array from one type to another without copying.