Redesigned the mixer output, so it is now always 32-bit PCM in any number of channels. Before it could be either 16-bit or 32-bit, mono or stereo. By doing this, it should be possible to play more than 2 channels in the future, e.g. sample files.
This change also requires the output agent to convert the 32-bit PCM to any format needed. Therefore I have improved the CoreAudio agent, to use any format and channels that Windows tell it needs instead of trying to ask for one. This should make NostalgicPlayer to play on Windows system which has been setup to e.g. 5.1 speaker setup.